fix: 修复 QColor::getCmykF 参数类型不匹配导致的编译错误
- 将 qreal 修改为 float 以对齐 Qt 6 API 签名
- 解决了 Windows 环境下 qreal(double) 到 float* 的隐式转换失败问题

+++ src/ui/MetaPanel.cpp
@@ -111,7 +111,7 @@ private:
         });
 
         menu.addAction("复制 cmyk格式", [this]() {
-            qreal c, m, y, k;
+            float c, m, y, k;
             m_color.getCmykF(&c, &m, &y, &k);
             QApplication::clipboard()->setText(QString("cmyk(%1%, %2%, %3%, %4%)")
                 .arg(qRound(c * 100)).arg(qRound(m * 100)).arg(qRound(y * 100)).arg(qRound(k * 100)));
